move rect adjacency matrix generator from GraphUtils. New test
