+++ src/test/java/com/roman_ks/maze/generator/adjacency/RectAdjacencyMatrixGeneratorTest.java
@@ -1,0 +1,89 @@
+package com.roman_ks.maze.generator.adjacency;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.roman_ks.maze.generator.utils.TestUtils;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import java.util.stream.Stream;
+
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.params.provider.Arguments.arguments;
+
+class RectAdjacencyMatrixGeneratorTest {
+
+    private static final ObjectMapper mapper = new ObjectMapper();
+
+    private static int[][] adjacencyMatrix2x2;
+    private static int[][] adjacencyMatrix3x3;
+    private static int[][] adjacencyMatrix4x4;
+    private static int[][] adjacencyMatrix10x5;
+
+    @BeforeAll
+    public static void loadMatrices() throws JsonProcessingException {
+        adjacencyMatrix2x2 = loadMatrix("matrices/2x2.json");
+        adjacencyMatrix3x3 = loadMatrix("matrices/3x3.json");
+        adjacencyMatrix4x4 = loadMatrix("matrices/4x4.json");
+        adjacencyMatrix10x5 = loadMatrix("matrices/10x5.json");
+    }
+
+    private static int[][] loadMatrix(String path) throws JsonProcessingException {
+        var json = TestUtils.readResource(path);
+        return mapper.readValue(json, int[][].class);
+    }
+
+    @ParameterizedTest
+    @MethodSource("provideAdjacentMatrixArgs")
+    void generateAdjacencyMatrixForRectangularMaze_square(int size, int[][] expectedMatrix) {
+        var adjacencyMatrixGenerator = new RectAdjacencyMatrixGenerator(size, size);
+        var matrix = adjacencyMatrixGenerator.generateAdjMatrix();
+
+        assertEquals(size * size, matrix.length);
+        assertEquals(size * size, matrix[0].length);
+
+        assertArrayEquals(expectedMatrix, matrix);
+    }
+
+    private static Stream<Arguments> provideAdjacentMatrixArgs() {
+        return Stream.of(
+                arguments(2, adjacencyMatrix2x2),
+                arguments(3, adjacencyMatrix3x3),
+                arguments(4, adjacencyMatrix4x4));
+    }
+
+    @Test
+    void generateAdjacencyMatrixForRectangularMaze_notSquare() {
+        var w = 10;
+        var h = 5;
+        var adjacencyMatrixGenerator = new RectAdjacencyMatrixGenerator(w, h);
+        var matrix = adjacencyMatrixGenerator.generateAdjMatrix();
+
+        assertEquals(w * h, matrix.length);
+        assertEquals(w * h, matrix[0].length);
+
+        assertArrayEquals(adjacencyMatrix10x5, matrix);
+    }
+
+    @Disabled
+    @ParameterizedTest
+    @ValueSource(ints = {2, 3, 4, 5})
+    void generateAdjacencyMatrixForRectangularMaze(int size) {
+        var adjacencyMatrixGenerator = new RectAdjacencyMatrixGenerator(size, size);
+        var matrix = adjacencyMatrixGenerator.generateAdjMatrix();
+
+        var mapper = new ObjectMapper();
+        try {
+            System.out.println(mapper.writeValueAsString(matrix));
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+        }
+    }
+
+}
