+++ src/test/java/com/roman_ks/maze/generator/utils/GraphUtilsTest.java
@@ -4,17 +4,14 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.roman_ks.maze.generator.model.Node;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.List;
 import java.util.stream.Stream;
 
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
@@ -36,24 +33,6 @@ class GraphUtilsTest {
     private static int[][] loadMatrix(String path) throws JsonProcessingException {
         var json = TestUtils.readResource(path);
         return mapper.readValue(json, int[][].class);
-    }
-
-    @ParameterizedTest
-    @MethodSource("provideAdjacentMatrixArgs")
-    void generateAdjacencyMatrixForRectangularMaze_square(int size, int[][] expectedMatrix) {
-        var matrix = GraphUtils.generateAdjacencyMatrixForRectangularMaze(size, size);
-
-        assertEquals(size * size, matrix.length);
-        assertEquals(size * size, matrix[0].length);
-
-        assertArrayEquals(expectedMatrix, matrix);
-    }
-
-    private static Stream<Arguments> provideAdjacentMatrixArgs() {
-        return Stream.of(
-                arguments(2, adjacencyMatrix2x2),
-                arguments(3, adjacencyMatrix3x3),
-                arguments(4, adjacencyMatrix4x4));
     }
 
     @ParameterizedTest
@@ -122,18 +101,4 @@ class GraphUtilsTest {
         }
     }
 
-    @Disabled
-    @ParameterizedTest
-    @ValueSource(ints = {2, 3, 4, 5})
-    void generateAdjacencyMatrixForRectangularMaze(int side) {
-        var matrix = GraphUtils.generateAdjacencyMatrixForRectangularMaze(side, side);
-
-        var mapper = new ObjectMapper();
-        try {
-            System.out.println(mapper.writeValueAsString(matrix));
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
-
-    }
 }
